+++ ui-test/ui-test.jsx
@@ -30,6 +30,7 @@ import {
     Pagination,
     SelectBox,
     TextField,
+    Chip,
 } from '../index.js';
 import {
     Layout, Content, Header
@@ -555,6 +556,26 @@ const Page = React.createClass({
             </div>
         );
 
+        const testChip = (
+            <div className="mdl-card mdl-shadow--2dp mdl-card--stretch">
+                <div className="mdl-card__title">
+                    <h4 className="mdl-card__title-text">Test Chip</h4>
+                </div>
+                <div className="mdl-card__content">
+                    <Chip
+                        label="clickable chip with customized icon"
+                        onClick={() => console.log('#1 chip clicked')}
+                        iconClassName={'bg-icon-denied tc-icon-white'}
+                        iconContent={'H'}
+                        className={'bg-chip-green tc-chip-blue'}
+                    />
+                    <Chip
+                        label="plain chip"
+                    />
+                </div>
+            </div>
+        );
+
         return (
             <Layout fixedHeader={true}>
                 <Header>
@@ -605,6 +626,7 @@ const Page = React.createClass({
                     <hr className="mdl-layout-spacer"/>
                     {testTextField}
                     <hr className="mdl-layout-spacer"/>
+                    {testChip}
                 </Content>
                 <footer className="mdl-mini-footer">
                     Footer