+++ src/Dialog/ConfirmationDialog.js
@@ -11,19 +11,14 @@ const ConfirmationDialog = React.createClass({
         className: React.PropTypes.string,
         modal: React.PropTypes.bool,
         size: React.PropTypes.string,
-        cancelButton: React.PropTypes.any.isRequired,
-        confirmButton: React.PropTypes.any.isRequired,
+        //a Confirmation Dialog should always have two buttons per spec:
+        // https://material.google.com/components/dialogs.html#dialogs-confirmation-dialogs
+        cancelButton: React.PropTypes.element.isRequired,
+        confirmButton: React.PropTypes.element.isRequired,
         title: React.PropTypes.node
     },
     // template rendering
     render() {
-        const buttons = [];
-        if (React.isValidElement(this.props.cancelButton)) {
-            buttons.push(this.props.cancelButton);
-        }
-        if (React.isValidElement(this.props.confirmButton)) {
-            buttons.push(this.props.confirmButton);
-        }
         // push data with formatted buttons to base dialog
         return (
             <BaseDialog
@@ -32,7 +27,10 @@ const ConfirmationDialog = React.createClass({
                 modal={this.props.modal}
                 size={this.props.size}
                 title={this.props.title}
-                buttonRow={buttons}
+                buttonRow={[
+                    this.props.cancelButton,
+                    this.props.confirmButton
+                ]}
             >
                 {this.props.children}
             </BaseDialog>