+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Changelog's Format](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- own `Chip` element
+
 ### Changed
 - Disabled buttons do not show their tooltip anymore
 - TODO: Describe Breaking CHANGE Checkbox and Switch